Fix the delay suites after their move under tests/
The delay suites were written when they lived in a top-level tests-colfee/
directory and were moved under tests/colfee/ so CI would run them. Their
relative requires were not adjusted with them, so every one of the six
resolved ../tests/Utils from one directory too shallow and CI failed at load
time with MODULE_NOT_FOUND before a single test ran. The storage-layout helper
resolved the repo root the same way and had the same fault.

Also completes MockArbitraryQuoteExitFeeController. It implements
IExitFeeController, which the delay work widened with the delay members, and
the mock was added on the fee line after this work branched — so it was never
extended and quietly compiled to an abstract contract with no bytecode. The
one suite that deploys it failed in its fixture. The delay members are
implemented inert: no delay, raw identities, which is the answer a controller
with the perimeter off returns, so the fee-side gate this mock exists to
exercise is unaffected.

tests/colfee now runs 107 passing, 0 failing.
