+++ tests/colfee/LenderExit.controllerPointer.failopen.test.js
@@ -42,8 +42,8 @@ const {
     getPriceFeeds,
     getSovryn,
     getSOV,
-} = require("../tests/Utils/initializer.js");
-const mutexUtils = require("../deployment/helpers/reentrancy/utils");
+} = require("../Utils/initializer.js");
+const mutexUtils = require("../../deployment/helpers/reentrancy/utils");
 
 const wei = web3.utils.toWei;
 const DELAY = 3600;
