+++ tests/colfee/LenderExit.delay.test.js
@@ -47,9 +47,9 @@ const {
     getPriceFeeds,
     getSovryn,
     getSOV,
-} = require("../tests/Utils/initializer.js");
-const mutexUtils = require("../deployment/helpers/reentrancy/utils");
-const { increaseTime } = require("../tests/Utils/Ethereum");
+} = require("../Utils/initializer.js");
+const mutexUtils = require("../../deployment/helpers/reentrancy/utils");
+const { increaseTime } = require("../Utils/Ethereum");
 
 const wei = web3.utils.toWei;
 const ZERO = "0x0000000000000000000000000000000000000000";
